+++ pystac/item.py
@@ -285,7 +285,9 @@ class Item(STACObject):
         else:
             return cast(Collection, collection_link.resolve_stac_object().target)
 
-    def to_dict(self, include_self_link: bool = True) -> Dict[str, Any]:
+    def to_dict(
+        self, include_self_link: bool = True, transform_hrefs: bool = True
+    ) -> Dict[str, Any]:
         links = self.links
         if not include_self_link:
             links = [x for x in links if x.rel != pystac.RelType.SELF]
@@ -303,7 +305,7 @@ class Item(STACObject):
             "id": self.id,
             "properties": self.properties,
             "geometry": self.geometry,
-            "links": [link.to_dict() for link in links],
+            "links": [link.to_dict(transform_href=transform_hrefs) for link in links],
             "assets": assets,
         }