+++ pystac/stac_object.py
@@ -428,12 +428,20 @@ class STACObject(ABC):
         raise NotImplementedError
 
     @abstractmethod
-    def to_dict(self, include_self_link: bool = True) -> Dict[str, Any]:
+    def to_dict(
+        self, include_self_link: bool = True, transform_hrefs: bool = True
+    ) -> Dict[str, Any]:
         """Generate a dictionary representing the JSON of this serialized object.
 
         Args:
             include_self_link : If True, the dict will contain a self link
                 to this object. If False, the self link will be omitted.
+            transform_hrefs: If True, transform the HREF of hierarchical links
+                based on the type of catalog this object belongs to (if any).
+                I.e. if this object belongs to a root catalog that is
+                RELATIVE_PUBLISHED or SELF_CONTAINED,
+                hierarchical link HREFs will be transformed to be relative to the
+                catalog root.
 
             dict: A serialization of the object that can be written out as JSON.
         """